ann: switch fully to SGD approach, working multi-layer!
Previously the implementation was split between a batch training method
and an SGD approach [1] where weights are updated for each input, to
just using the SGD approach to simplify things. Here we resolve that by
eliminating the partial batching support, which we can bring back in
generically later on.

This in turn gets the multi-layer test cases working! Probably indicating
that the split approach was causing subtle issues before.

[1] https://www.geeksforgeeks.org/ml-mini-batch-gradient-descent-with-python/