neural: add single-layer bool test cases
This helps validate the network at a simpler level, prompting us to
refactor out the test cases for re-use.

In this change, we are also see that we can reduce the number of epochs
for these training processes and speed up the test cases a bit while
reducing the volume of log output.