neural: leverage more linear algebra in backprop
Also expands the lin package to add additional functionality supporting
this change.

The motivation for this commit was to try to make the backprop function
easier to read through, and I found the outer for loop to be confusing,
as it was iterating over nodes in the layer only, not being used
everywhere, and potentially prone to mental confusion with the
iterations over layers that lives outside of this function.

Not everything was converted to a purely linear algebraic form in order
to keep some of the more complex pieces still clear and spelled out in
front of us, though that could be iterated on further in the future.